refactor(modal dialog): Refactored modal dialog system with improved widget management and li...
# Project State

## Current Focus
Refactored modal dialog system with improved widget management and lifecycle handling

## Context
The changes address recent refactoring of the modal dialog system to better manage widget lifetimes and improve type safety in the terminal engine.

## Completed
- [x] Added proper lifetime annotations to `HelpOverlay` and `ModalDemoApp` structs
- [x] Fixed potential integer overflow in theme switcher's cell index calculation
- [x] Updated Cargo.lock with dependency version changes

## In Progress
- [ ] No active work in progress shown in these changes

## Blockers
- None identified in this commit

## Next Steps
1. Verify modal dialog behavior in the demo examples
2. Test widget lifecycle management with the updated implementations
---
category: refactor
scope: modal dialog
---

+++ tools/tui/dracon-terminal-engine/examples/theme_switcher.rs
@@ -181,7 +181,7 @@ impl Widget for ThemeHeader {
 
         for y in 0..area.height {
             for x in 0..area.width {
-                let idx = (y * area.width + x) as usize;
+                let idx = y as usize * area.width as usize + x as usize;
                 if idx < plane.cells.len() {
                     plane.cells[idx] = Cell {
                         char: ' ',
